Solid engine (#15)
* Fixes Scene constructor and the possible memory fault

* Serialization fix

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* NAMSC-41 event tree ( scenery objects not implemented yet) (#12)

* NAMSC-41 - added event tree part 1

* NAMSC-41 - fixed conflicts ( not working )

* NAMSC-41 - connected event tree to properties ( todo: sceneryobjects in event tree )

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Serialization fix

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Fixes GetType ambiguity, turns Scenery's maps into lists and provides helpers for them. Adds Novel clear method. Changes instanized templates to true templates. Merges with changes already uploaded to the master. Minor bugfixes and QoL

* Implements the Novel Engine that can handle EventDialogue. Lacks cps feature

* Hotfix and cps foundations

* Adds cps, more performance on scaling.

* Serialization fix

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Fixes GetType ambiguity, turns Scenery's maps into lists and provides helpers for them. Adds Novel clear method. Changes instanized templates to true templates. Merges with changes already uploaded to the master. Minor bugfixes and QoL

* Implements the Novel Engine that can handle EventDialogue. Lacks cps feature

* Hotfix and cps foundations

* Adds cps, more performance on scaling.

* Finished

+++ NAMSC_lib/src/Novel/Data/Scene.cpp
@@ -147,9 +147,9 @@ bool Scene::operator==(const Scene& obj) const noexcept
 void Scene::serializableLoad(QDataStream& dataStream)
 {
     dataStream >> name >> chapterName_ >> scenery;
-    size_t size;
+    uint size;
     dataStream >> size;
-    for (size_t i = 0u; i != size; ++i)
+    for (uint i = 0u; i != size; ++i)
     {
         //TODO: Create an abstract factory for this
         Event* event = nullptr;
